feat(`headerfs`): separate file-system dependencies in `headerStore`
